+++ src/NJsonSchema/JsonSchemaResolver.cs
@@ -13,24 +13,19 @@ using NJsonSchema.Generation;
 namespace NJsonSchema
 {
     /// <summary>Manager which resolves types to schemas and appends missing schemas to the root object.</summary>
-    public class JsonSchemaResolver
+    public class JsonSchemaResolver : JsonSchemaAppender
     {
         private readonly Dictionary<string, JsonSchema4> _mappings = new Dictionary<string, JsonSchema4>();
         private readonly JsonSchemaGeneratorSettings _settings;
-
-        private JsonSchema4 RootSchema => (JsonSchema4)RootObject;
 
         /// <summary>Initializes a new instance of the <see cref="JsonSchemaResolver" /> class.</summary>
         /// <param name="rootObject">The root schema.</param>
         /// <param name="settings">The settings.</param>
         public JsonSchemaResolver(object rootObject, JsonSchemaGeneratorSettings settings)
+            : base(rootObject, settings.TypeNameGenerator)
         {
             _settings = settings;
-            RootObject = rootObject; 
         }
-
-        /// <summary>Gets the root object.</summary>
-        public object RootObject { get; }
 
         /// <summary>Determines whether the specified type has a schema.</summary>
         /// <param name="type">The type.</param>
@@ -64,28 +59,6 @@ namespace NJsonSchema
                 AppendSchema(schema, _settings.SchemaNameGenerator.Generate(type));
 
             _mappings.Add(GetKey(type, isIntegerEnumeration), schema);
-        }
-
-        /// <summary>Appends the schema to the root object.</summary>
-        /// <param name="schema">The schema to append.</param>
-        /// <param name="typeNameHint">The type name hint.</param>
-        /// <exception cref="ArgumentNullException"><paramref name="schema"/> is <see langword="null"/></exception>
-        /// <exception cref="ArgumentException">The root schema cannot be appended.</exception>
-        public virtual void AppendSchema(JsonSchema4 schema, string typeNameHint)
-        {
-            if (schema == null)
-                throw new ArgumentNullException(nameof(schema));
-            if (schema == RootObject)
-                throw new ArgumentException("The root schema cannot be appended.");
-
-            if (!RootSchema.Definitions.Values.Contains(schema))
-            {
-                var typeName = _settings.TypeNameGenerator.Generate(schema, typeNameHint, RootSchema.Definitions.Keys);
-                if (!string.IsNullOrEmpty(typeName) && !RootSchema.Definitions.ContainsKey(typeName))
-                    RootSchema.Definitions[typeName] = schema;
-                else
-                    RootSchema.Definitions["ref_" + Guid.NewGuid().ToString().Replace("-", "_")] = schema;
-            }
         }
 
         /// <summary>Gets all the schemas.</summary>
